Organize mocks into a directory in test/
