+++ platform/kube/BUILD
@@ -44,7 +44,7 @@ go_test(
     library = ":go_default_library",
     deps = [
         "//model:go_default_library",
-        "//test:go_default_library",
+        "//test/mock:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
         "@io_k8s_client_go//pkg/api/v1:go_default_library",
